Flatten the lib
